More bug fixes
