--- MultiRPC/Assets/Icons.html
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<html>
-  <head>
-    <title>Browse svg images</title>
-  </head>
-  <body>Images in file: .\Icons<br /><img src="AddIcon.svg" title="AddIcon" height="128" width="128" /><img src="BrowserIcon.svg" title="BrowserIcon" height="128" width="128" /><img src="CreditsIcon.svg" title="CreditsIcon" height="128" width="128" /><img src="CustomIcon.svg" title="CustomIcon" height="128" width="128" /><img src="DeleteIcon.svg" title="DeleteIcon" height="128" width="128" /><img src="DiscordColourIcon.svg" title="DiscordColourIcon" height="128" width="128" /><img src="DiscordIcon.svg" title="DiscordIcon" height="128" width="128" /><img src="DownloadIcon.svg" title="DownloadIcon" height="128" width="128" /><img src="GithubIcon.svg" title="GithubIcon" height="128" width="128" /><img src="HeartIcon.svg" title="HeartIcon" height="128" width="128" /><img src="HelpIcon.svg" title="HelpIcon" height="128" width="128" /><img src="LogsIcon.svg" title="LogsIcon" height="128" width="128" /><img src="MultiRPCLogo.svg" title="MultiRPCLogo" height="128" width="128" /><img src="PencilIcon.svg" title="PencilIcon" height="128" width="128" /><img src="SettingsIcon.svg" title="SettingsIcon" height="128" width="128" /><img src="ShareIcon.svg" title="ShareIcon" height="128" width="128" /><img src="ShieldIcon.svg" title="ShieldIcon" height="128" width="128" /></body>
-</html>
