+++ MultiRPC/Assets/Icons.xaml
@@ -411,16 +411,6 @@
         </DrawingGroup>
     </DrawingGroup>
     <DrawingImage x:Key="SettingsIconDrawingImage" Drawing="{StaticResource SettingsIconDrawingGroup}" />
-    <Geometry x:Key="MultiRPCLogoGeometry">F1 M1000,1000z M0,0z M617.94,676C617.94,676 602.7,657.79 589.94,642.12 645.42,626.45 666.59,592.12 666.59,592.12A242.81,242.81,0,0,1,617.89,617.12A304.4,304.4,0,0,1,556.49,635.33A296.72,296.72,0,0,1,446.8,634.9A322.36,322.36,0,0,1,384.55,616.69A248.08,248.08,0,0,1,353.63,602.29C352.36,601.45 351.09,601.02 349.82,600.18 348.98,599.75 348.55,599.33 348.13,599.33 340.51,595.09 336.27,592.13 336.27,592.13 336.27,592.13 356.6,625.59 410.38,641.68 397.68,657.77 382.01,676.4 382.01,676.4 288.42,673.4 252.84,612.46 252.84,612.46 252.84,477.37 313.84,367.68 313.84,367.68 374.84,322.37 432.41,323.68 432.41,323.68L436.65,328.76C360.42,350.36 325.65,383.81 325.65,383.81 325.65,383.81 334.97,378.73 350.65,371.96 396,352 432,346.91 446.85,345.22A41.11,41.11,0,0,1,454.05,344.37A358.13,358.13,0,0,1,539.6,343.52A352.22,352.22,0,0,1,667.07,383.75C667.07,383.75,633.61,352,561.62,330.39L567.55,323.62C567.55,323.62 625.55,322.35 686.12,367.62 686.12,367.62 747.12,477.31 747.12,612.4 747.12,611.97 711.54,672.96 617.95,675.92z M577.59,479.89C553.45,479.89 534.4,500.64 534.4,526.47 534.4,552.3 553.88,573.05 577.59,573.05 601.73,573.05 620.79,552.3 620.79,526.47 620.79,500.64 601.31,479.89 577.59,479.89z M423.02,479.89C398.88,479.89 379.82,500.64 379.82,526.47 379.82,552.3 399.3,573.05 423.02,573.05 447.16,573.05 466.22,552.3 466.22,526.47 466.64,500.64 447.16,479.89 423,479.89z</Geometry>
-    <DrawingGroup x:Key="MultiRPCLogoDrawingGroup" ClipGeometry="M0,0 V1000 H1000 V0 H0 Z">
-        <GeometryDrawing Brush="{DynamicResource BackgroundGradient}">
-            <GeometryDrawing.Geometry>
-                <RectangleGeometry RadiusX="0" RadiusY="0" Rect="0,0,1000,1000" />
-            </GeometryDrawing.Geometry>
-        </GeometryDrawing>
-        <GeometryDrawing Brush="#FFFFFFFF" Geometry="{StaticResource MultiRPCLogoGeometry}" />
-    </DrawingGroup>
-    <DrawingImage x:Key="MultiRPCLogoDrawingImage" Drawing="{StaticResource MultiRPCLogoDrawingGroup}" />
     <Geometry x:Key="LogsIconGeometry1">F1 M1000,1000z M0,0z M184,314.1L544.4,314.1 544.4,346.9 184,346.9 184,314.1 184,314.1z</Geometry>
     <Geometry x:Key="LogsIconGeometry2">F1 M1000,1000z M0,0z M184,445.3L394.2,445.3 394.2,478.1 184,478.1 184,445.3z</Geometry>
     <Geometry x:Key="LogsIconGeometry3">F1 M1000,1000z M0,0z M184,576.5L334.2,576.5 334.2,609.3 184,609.3 184,576.5z</Geometry>
@@ -514,13 +504,6 @@
         </DrawingGroup>
     </DrawingGroup>
     <DrawingImage x:Key="DebugIconSelectedDrawingImage" Drawing="{StaticResource DebugIconSelectedDrawingGroup}" />
-
-    <LinearGradientBrush x:Key="BackgroundGradient" StartPoint="1091.51,158.49" EndPoint="-91.51,841.51"
-                         MappingMode="Absolute" SpreadMethod="Pad" Opacity="1">
-        <GradientStop Color="#FF00FFED" Offset="0" />
-        <GradientStop Color="#FF0077E9" Offset="1" />
-    </LinearGradientBrush>
-
     <DrawingGroup x:Key="ThemeIconSelectedDrawingGroup" ClipGeometry="M0,0 V1000 H1000 V0 H0 Z">
         <GeometryDrawing Brush="{DynamicResource AccentColour5SCBrush}">
             <GeometryDrawing.Geometry>
