+++ MultiRPC/GUI/Controls/TabPage.xaml
@@ -1,0 +1,26 @@
+﻿<Page x:Class="MultiRPC.GUI.Controls.TabPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:local="clr-namespace:MultiRPC.GUI.Controls"
+      mc:Ignorable="d"
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="TabPage">
+
+    <Grid Background="{DynamicResource AccentColour1SCBrush}">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="*" />
+        </Grid.RowDefinitions>
+        <Grid Margin="10">
+            <StackPanel Orientation="Horizontal" Name="spTabContainer" />
+        </Grid>
+        <Border Name="mask" Style="{DynamicResource maskStyle}" Grid.Row="1" />
+        <Frame Name="frmContent" Grid.Row="1">
+            <Frame.OpacityMask>
+                <VisualBrush Visual="{Binding ElementName=mask}" />
+            </Frame.OpacityMask>
+        </Frame>
+    </Grid>
+</Page>
