+++ MultiRPC/GUI/MessageBox/CustomMessageBox.cs
@@ -22,16 +22,33 @@ namespace MultiRPC.GUI
 
         private static async Task<MessageBoxResult> _Show(string messageBoxText, string messageBoxTitle,
             MessageBoxButton messageBoxButton, MessageBoxImage messageBoxImage,
-            MessageBoxResult messageBoxResult = MessageBoxResult.None, Window Ownerwindow = null)
+            MessageBoxResult messageBoxResult = MessageBoxResult.None, Window ownerWindow = null)
         {
             await Application.Current.Dispatcher.InvokeAsync(() =>
             {
                 var tick = DateTime.Now.Ticks;
                 var page = new MessageBoxPage(messageBoxText, messageBoxButton, messageBoxImage, tick);
+
+                switch (messageBoxButton)
+                {
+                    case MessageBoxButton.OKCancel:
+                        page.btnOk.IsDefault = true;
+                        break;
+                    case MessageBoxButton.OK:
+                        page.btnOk.IsDefault = true;
+                        break;
+                    case MessageBoxButton.YesNoCancel:
+                        page.btnYes.IsDefault = true;
+                        break;
+                    case MessageBoxButton.YesNo:
+                        page.btnYes.IsDefault = true;
+                        break;
+                }
+
                 var window = new CustomMessageBox(page, messageBoxTitle)
                 {
                     WindowID = tick,
-                    Owner = Ownerwindow
+                    Owner = ownerWindow
                 };
                 if (messageBoxButton == MessageBoxButton.YesNo)
                     window.btnClose.IsEnabled = false;
@@ -51,6 +68,8 @@ namespace MultiRPC.GUI
                     return MessageBoxResult.OK;
                 case MessageBoxButton.YesNoCancel:
                     return MessageBoxResult.Cancel;
+                case MessageBoxButton.YesNo:
+                    return MessageBoxResult.Yes;
                 default:
                     return defaultMessageBoxResult;
             }
@@ -59,27 +78,27 @@ namespace MultiRPC.GUI
         public static async Task<MessageBoxResult> Show(string messageBoxText, Window window = null)
         {
             return await _Show(messageBoxText, "MultiRPC", MessageBoxButton.OK, MessageBoxImage.None,
-                Ownerwindow: window);
+                ownerWindow: window);
         }
 
         public static async Task<MessageBoxResult> Show(string messageBoxText, string messageBoxTitle,
             Window window = null)
         {
             return await _Show(messageBoxText, messageBoxTitle, MessageBoxButton.OK, MessageBoxImage.None,
-                Ownerwindow: window);
+                ownerWindow: window);
         }
 
         public static async Task<MessageBoxResult> Show(string messageBoxText, string messageBoxTitle,
             MessageBoxButton messageBoxButton, Window window = null)
         {
             return await _Show(messageBoxText, messageBoxTitle, messageBoxButton, MessageBoxImage.None,
-                Ownerwindow: window);
+                ownerWindow: window);
         }
 
         public static async Task<MessageBoxResult> Show(string messageBoxText, string messageBoxTitle,
             MessageBoxButton messageBoxButton, MessageBoxImage messageBoxImage, Window window = null)
         {
-            return await _Show(messageBoxText, messageBoxTitle, messageBoxButton, messageBoxImage, Ownerwindow: window);
+            return await _Show(messageBoxText, messageBoxTitle, messageBoxButton, messageBoxImage, ownerWindow: window);
         }
 
         public static async Task<MessageBoxResult> Show(string messageBoxText, string messageBoxTitle,
