+++ MultiRPC/GUI/Pages/CustomPage.xaml
@@ -37,8 +37,7 @@
         </Grid.ColumnDefinitions>
 
         <Rectangle Grid.ColumnSpan="2" Grid.Row="1" Fill="{DynamicResource AccentColour1SCBrush}" />
-        <Border CornerRadius="15,15,0,0" Grid.ColumnSpan="2" Grid.Row="1"
-                Background="{DynamicResource AccentColour2SCBrush}" />
+        <Border Style="{DynamicResource maskStyle}" Grid.ColumnSpan="2" Grid.Row="1" />
 
         <Grid Grid.ColumnSpan="2" HorizontalAlignment="Stretch" Background="{DynamicResource AccentColour1SCBrush}">
             <Grid.RowDefinitions>
@@ -48,9 +47,8 @@
             <ToolBar Margin="5,0,0,0" Background="{DynamicResource AccentColour1SCBrush}"
                      Foreground="{DynamicResource TextColourSCBrush}" x:Name="tbProfiles" />
             <StackPanel Margin="0,5" Grid.Row="1" Orientation="Horizontal">
-                <TextBlock FontWeight="SemiBold" FontSize="16"
-                           Text="If you see this then the program hasn't click the first profile (oops)"
-                           Name="tblProfileName" Margin="10,0,5,0" />
+                <TextBlock FontWeight="SemiBold" Style="{DynamicResource SubTitle}"
+                           Name="tblProfileName" Margin="10,0,5,0" VerticalAlignment="Center" />
                 <Rectangle Margin="0,2" Fill="{DynamicResource AccentColour3SCBrush}" Width="0.25"
                            VerticalAlignment="Stretch" />
                 <Image Name="imgProfileEdit" Opacity="0.6" Source="{DynamicResource PencilIconDrawingImage}"
