+++ MultiRPC/GUI/Pages/Theme Pages/InstalledThemes.xaml
@@ -1,0 +1,20 @@
+﻿<Page x:Class="MultiRPC.GUI.Pages.InstalledThemes"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:local="clr-namespace:MultiRPC.GUI.Pages"
+      mc:Ignorable="d"
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="InstalledThemes">
+
+    <ScrollViewer>
+        <StackPanel Margin="10">
+            <StackPanel Margin="0,0,0,10" HorizontalAlignment="Left" VerticalAlignment="Top" Orientation="Horizontal">
+                <Button Name="btnAddTheme" Click="BtnAddTheme_OnClick" />
+                <Button Margin="5,0,0,0" Name="btnAddAndApplyTheme" Click="BtnAddAndApplyTheme_OnClick" />
+            </StackPanel>
+            <WrapPanel Name="wpInstalledThemes" HorizontalAlignment="Center" MaxWidth="735" />
+        </StackPanel>
+    </ScrollViewer>
+</Page>
