+++ MultiRPC/GUI/Pages/Theme Pages/MasterThemeEditorPage.xaml
@@ -1,0 +1,12 @@
+﻿<Page x:Class="MultiRPC.GUI.Pages.MasterThemeEditorPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:local="clr-namespace:MultiRPC.GUI.Pages"
+      mc:Ignorable="d"
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="MasterThemeEditorPage">
+
+    <Frame Name="frmContent" />
+</Page>
