+++ MultiRPC/GUI/Pages/Theme Pages/ThemeEditorPage.xaml
@@ -24,14 +24,9 @@
                         <RowDefinition Height="Auto" />
                         <RowDefinition Height="Auto" />
                         <RowDefinition Height="Auto" />
-                        <RowDefinition Height="Auto" />
                     </Grid.RowDefinitions>
                     <TextBlock Name="tblMakeTheme" HorizontalAlignment="Center" VerticalAlignment="Center"
-                               FontWeight="Bold" FontSize="25" />
-                    <StackPanel Margin="0,0,0,10" HorizontalAlignment="Left" VerticalAlignment="Top">
-                        <Button Name="btnAddTheme" Click="BtnAddTheme_OnClick" />
-                        <Button Margin="0,5,0,0" Name="btnAddAndApplyTheme" Click="BtnAddAndApplyTheme_OnClick" />
-                    </StackPanel>
+                               FontWeight="Bold" Style="{DynamicResource Title}" Margin="0,0,0,10" />
                     <StackPanel Grid.Row="1" HorizontalAlignment="Stretch">
                         <StackPanel Orientation="Horizontal" HorizontalAlignment="Center">
                             <xctk:ColorCanvas Name="ccvcolourPicker"
@@ -410,11 +405,6 @@
                             IsEnabled="False" />
                     <Button Margin="5,0,0,0" Name="btnResetTheme" Click="BtnResetTheme_OnClick" />
                 </StackPanel>
-            </StackPanel>
-
-            <StackPanel Grid.Row="3" Margin="0,10">
-                <TextBlock Name="tblInstalledThemes" FontWeight="SemiBold" FontSize="18" Margin="0,0,0,15" />
-                <WrapPanel Name="wpInstalledThemes" HorizontalAlignment="Center" MaxWidth="735" />
             </StackPanel>
         </Grid>
     </ScrollViewer>
